renaming files. _test.cc and _main.cc added
